collect deprecated info of api
