+++ idea-plugin/src/main/kotlin/com/itangcent/idea/plugin/actions/MarkdownExportAction.kt
@@ -3,6 +3,7 @@ package com.itangcent.idea.plugin.actions
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.components.ServiceManager
 import com.intellij.openapi.project.Project
+import com.itangcent.idea.plugin.api.export.DefaultDocParseHelper
 import com.itangcent.idea.plugin.api.export.DocParseHelper
 import com.itangcent.idea.plugin.api.export.EasyApiConfigReader
 import com.itangcent.idea.plugin.api.export.markdown.MarkdownApiExporter
@@ -15,9 +16,6 @@ import com.itangcent.intellij.extend.guice.with
 import com.itangcent.intellij.file.DefaultLocalFileRepository
 import com.itangcent.intellij.file.LocalFileRepository
 import com.itangcent.intellij.psi.ClassRuleConfig
-import com.itangcent.intellij.psi.DefaultClassRuleConfig
-import com.itangcent.intellij.setting.ReadOnlySettingManager
-import com.itangcent.intellij.setting.SettingManager
 
 class MarkdownExportAction : ApiExportAction("Export Markdown") {
 
@@ -27,7 +25,7 @@ class MarkdownExportAction : ApiExportAction("Export Markdown") {
         builder.bind(LocalFileRepository::class) { it.with(DefaultLocalFileRepository::class).singleton() }
         builder.bind(SettingBinder::class) { it.toInstance(ServiceManager.getService(SettingBinder::class.java)) }
         builder.bind(MarkdownApiExporter::class) { it.singleton() }
-        builder.bind(DocParseHelper::class) { it.singleton() }
+        builder.bind(DocParseHelper::class) { it.with(DefaultDocParseHelper::class).singleton() }
         builder.bind(ClassRuleConfig::class) { it.with(RecommendClassRuleConfig::class).singleton() }
         builder.bind(ConfigReader::class) { it.with(EasyApiConfigReader::class).singleton() }
