+++ idea-plugin/src/main/kotlin/com/itangcent/idea/plugin/actions/PostmanExportAction.kt
@@ -3,6 +3,7 @@ package com.itangcent.idea.plugin.actions
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.components.ServiceManager
 import com.intellij.openapi.project.Project
+import com.itangcent.idea.plugin.api.export.DefaultDocParseHelper
 import com.itangcent.idea.plugin.api.export.DocParseHelper
 import com.itangcent.idea.plugin.api.export.postman.*
 import com.itangcent.idea.plugin.settings.SettingBinder
@@ -28,7 +29,7 @@ class PostmanExportAction : ApiExportAction("Export Postman") {
         builder.bind(PostmanApiExporter::class) { it.singleton() }
         builder.bind(PostmanFormatter::class) { it.singleton() }
         builder.bindInstance(HttpClient::class, HttpClients.createDefault())
-        builder.bind(DocParseHelper::class) { it.singleton() }
+        builder.bind(DocParseHelper::class) { it.with(DefaultDocParseHelper::class).singleton() }
         builder.bind(ClassRuleConfig::class) { it.with(RecommendClassRuleConfig::class).singleton() }
         builder.bind(ConfigReader::class) { it.with(PostmanConfigReader::class).singleton() }
